pyproject.toml
[tool.poetry]
name = "jararaca"
version = "0.3.11a4"
description = "A simple and fast API framework for Python"
authors = ["Lucas S <me@luscasleo.dev>"]
readme = "README.md"
packages = [{ include = "jararaca", from = "src" }]
include = ["pyproject.toml", "README.md", "LICENSE", "docs"]
repository = "https://github.com/LuscasLeo/jararaca"


[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.6"
uvloop = "^0.20.0"
fastapi = "^0.113.0"
aio-pika = "^9.4.3"
sqlalchemy = "^2.0.34"
croniter = "^3.0.3"
redis = "^5.0.8"
websockets = "^13.0.1"
opentelemetry-exporter-otlp-proto-http = { version = "^1.27.0", optional = true }
opentelemetry-api = { version = "^1.27.0", optional = true }
opentelemetry-sdk = { version = "^1.27.0", optional = true }
opentelemetry-distro = { version = "^0.49b2", optional = true }
opentelemetry-exporter-otlp = { version = "^1.27.0", optional = true }
types-croniter = "^3.0.3.20240731"
types-redis = "^4.6.0.20240903"
mako = "^1.3.5"
watchdog = { version = "^3.0.0", optional = true }
frozendict = "^2.4.6"


[tool.poetry.extras]
opentelemetry = [
    "opentelemetry-exporter-otlp-proto-http",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-distro",
    "opentelemetry-exporter-otlp",
]
http = ["httptools", "httpx"]
docs = ["mkdocs-material"]
watch = ["watchdog"]

[tool.poetry.group.lint.dependencies]
mypy = "^1.11.2"
black = "^24.8.0"
isort = "^5.13.2"
pre-commit = "^3.8.0"
autoflake = "^2.3.1"
types-croniter = "^3.0.3.20240731"
types-redis = "^4.6.0.20240903"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.5.34"
mkdocs-mermaid2-plugin = "^1.2.1"


[tool.poetry.group.dev.dependencies]
httptools = "^0.6.1"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
strict = true
# mypy_path = "src"


# JARARACA CONFIG
[tool.poetry.scripts]
jararaca = "jararaca.cli:cli"

[[tool.mypy.overrides]]
module = "mako.*"
ignore_missing_imports = true
